refactor(feeder): move get and buildQueryString methods to request.go file
